Make json serializable workflow trait so code can be shared
